+++ graphannis/src/annis/db/aql/disjunction.rs
@@ -36,17 +36,23 @@ impl Disjunction {
         None
     }
 
-    /// Return the variable name for a given position in the match output list.
-    ///
-    /// Optional nodes that are not part of the output are ignored. If there are
-    /// no optional nodes, this corresponds to the index of the node in the
-    /// query.
-    pub(crate) fn get_variable_by_pos(&self, pos: usize) -> Option<String> {
+    /// Return the variable name for a node number. The node number is the
+    /// position of an AQL query node in the disjunction.
+    pub(crate) fn get_variable_by_node_nr(&self, node_nr: usize) -> Option<String> {
         for alt in &self.alternatives {
-            if let Some(var) = alt.get_variable_by_pos(pos) {
+            if let Some(var) = alt.get_variable_by_node_nr(node_nr) {
                 return Some(var);
             }
         }
         None
+    }
+
+    pub(crate) fn is_included_in_output(&self, variable: &str) -> bool {
+        for alt in &self.alternatives {
+            if alt.is_included_in_output(variable) {
+                return true;
+            }
+        }
+        false
     }
 }